modificando tipo de chave primaria das tabelas

+++ src/main/resources/db/migration/V5__create_table_livro-emprestimo.sql
@@ -1,6 +1,6 @@
 CREATE TABLE livro_emprestimo (
-                                  emprestimo_id INT,
-                                  livro_id INT,
+                                  emprestimo_id BIGINT,
+                                  livro_id BIGINT,
                                   PRIMARY KEY (emprestimo_id, livro_id),
                                   FOREIGN KEY (emprestimo_id) REFERENCES emprestimo(id),
                                   FOREIGN KEY (livro_id) REFERENCES livro(id)
